Deleting from an unknown table raises error
If the table exists then we deletes are idempotent

+++ moto/dynamodb2/responses.py
@@ -418,6 +418,11 @@ class DynamoHandler(BaseResponse):
         name = self.body['TableName']
         keys = self.body['Key']
         return_values = self.body.get('ReturnValues', '')
+        table = dynamodb_backend2.get_table(name)
+        if not table:
+            er = 'com.amazonaws.dynamodb.v20120810#ConditionalCheckFailedException'
+            return self.error(er)
+
         item = dynamodb_backend2.delete_item(name, keys)
         if item and return_values == 'ALL_OLD':
             item_dict = item.to_json()
